refactor(httpd): share JSON/API error envelope helpers
